CDE-54 changes made
